+++ springql-core/src/api.rs
@@ -13,6 +13,7 @@ pub use crate::{
         spring_config::*,
         spring_pipeline::SpringPipeline,
         spring_sink_row::SpringSinkRow,
+        spring_source_row::SpringSourceRow,
         SpringConfig,
     },
     stream_engine::{